requirements.txt
async-timeout==4.0.3
blinker==1.7.0
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.7.0
defusedxml==0.7.1
discord-webhook==1.3.0
Flask==3.0.1
Flask-HTTPAuth==4.8.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
MarkupSafe==2.1.4
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pycparser==2.21
python-dateutil==2.8.2
rauth==0.7.3
realtime==1.0.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
tinycss2==1.2.1
tweepy==4.14.0
typing_extensions==4.9.0
urllib3==2.1.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
